Change init() to isChecked
